[Core, CoreBundle] move event based Payment Provider Resolver to Bundle
