+++ src/CoreShop/Bundle/CoreBundle/Resources/config/services/payment.yml
@@ -11,7 +11,7 @@ services:
             - '@coreshop.context.store'
 
     coreshop.payment.event_based_payment_provider_resolver:
-        class: CoreShop\Component\Core\Payment\Resolver\EventBasedPaymentProviderResolver
+        class: CoreShop\Bundle\CoreBundle\Payment\Resolver\EventBasedPaymentProviderResolver
         decoration_priority: 5
         decorates: 'coreshop.payment.payment_provider_resolver'
         arguments:
